+++ api/exceptions/empty_password_error.py
