+++ api/models/responses/exercises.py
@@ -1,0 +1,5 @@
+from pydantic.main import BaseModel
+
+
+class Exercises(BaseModel):
+    exercises: list
